Feature/analyze url

+++ backend-api/app/controllers/controller.js
@@ -4,8 +4,8 @@ const moment = require("moment");
 const prependFile = require("prepend-file");
 const AWS = require("aws-sdk");
 const db = require("../models");
-const { checkIsUrl } = require("./inspectionmethods");
-const InspectLink = db.inspected_links;
+const { checkIsUrl } = require("./inspection.controller");
+const InspectLinks = db.inspected_links;
 const { ObjectId } = require("mongodb");
 
 // var credentials = new AWS.SharedIniFileCredentials({ profile: "default" });
@@ -32,8 +32,8 @@ exports.inspectLink = (req, res) => {
 
   var url = req.body.inspectURL;
 
-  /* --------------------- Creating an InspectLink object --------------------- */
-  var inspectLink = {
+  /* --------------------- Creating an InspectLinks object --------------------- */
+  var inspectedLink = {
     processed_url: "",
     original_url: url,
     status: "processing", // status as "processing" to indicate that the processing is still ongoing
@@ -41,6 +41,8 @@ exports.inspectLink = (req, res) => {
     image: "",
     domain_age: null,
     flag_points: 0,
+    registrar_abuse_contact: "",
+    toFlag: null
   };
 
   /* -------------------------------------------------------------------------- */
@@ -59,7 +61,7 @@ exports.inspectLink = (req, res) => {
   /*                  Create new Worker Thread to inspect link                  */
   /* -------------------------------------------------------------------------- */
   const worker = new Worker("./app/controllers/inspectionWorker.js", {
-    workerData: { url: url, inspectLink: inspectLink },
+    workerData: { url: url, inspectedLink: inspectedLink },
   });
 
   worker.on("message", (message) => {
@@ -69,8 +71,8 @@ exports.inspectLink = (req, res) => {
     } else if (message[0] == "flag") {
       prependLine(fileName, message[1]);
     } else if (message[0] == "termination") {
-      inspectLink = message[1];
-      inspectLink._id = ObjectId(inspectLink._id);
+      inspectedLink = message[1];
+      inspectedLink._id = ObjectId(inspectedLink._id);
     }
   });
 
@@ -79,39 +81,45 @@ exports.inspectLink = (req, res) => {
   });
 
   worker.on("exit", (exitCode) => {
-    writeLine(logger, "Link inspection completed.");
-    prependLine(
-      fileName,
-      "--------------------- Flags ----------------------"
-    ).then(() => {
-      //configuring parameters
-      var params = {
-        Bucket: "scam-sword-link-inspection-reports",
-        Body: fs.createReadStream(fileName),
-        Key: fileName,
-      };
-
-      s3.upload(params, function (err, data) {
-        //handle error
-        if (err) {
-          console.log("Error", err);
-        }
-
-        //success
-        if (data) {
-          console.log("Uploaded in:", data.Location);
-          fs.unlinkSync(fileName);
-          InspectLink.findOne(
-            { _id: inspectLink._id },
-            function (error, inspectLink) {
-              if (error) console.log(error);
-              inspectLink.report = data.Location;
-              inspectLink.save();
-            }
-          );
-        }
+    console.log(exitCode);
+    if (exitCode == 0) {
+      writeLine(logger, "Link inspection completed.");
+      prependLine(
+        fileName,
+        "--------------------- Flags ----------------------"
+      ).then(() => {
+        //configuring parameters
+        var params = {
+          Bucket: "scam-sword-link-inspection-reports",
+          Body: fs.createReadStream(fileName),
+          Key: fileName,
+        };
+  
+        s3.upload(params, function (err, data) {
+          //handle error
+          if (err) {
+            console.log("Error", err);
+          }
+  
+          //success
+          if (data) {
+            console.log(inspectedLink._id);
+            console.log("Uploaded in:", data.Location);
+            fs.unlinkSync(fileName);
+            InspectLinks.findOne(
+              { _id: inspectedLink._id },
+              function (error, record) {
+                if (error) console.log(error);
+                else {
+                  record.report = data.Location;
+                  record.save();
+                }
+              }
+            );
+          }
+        });
       });
-    });
+    }
   });
 
   res.send({
